% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_featdata.R
\name{add_featdata}
\alias{add_featdata}
\alias{add_featdata.profile.data}
\title{Add featdata}
\usage{
add_featdata(P, ...)

\method{add_featdata}{profile.data}(P, P_new, ...)
}
\arguments{
\item{P}{profile.data}

\item{...}{additional parameters}

\item{P_new}{profile.data with new featdata columns}
}
\description{
Add featdata
}
\section{Methods (by class)}{
\itemize{
\item \code{profile.data}: Add featdata to profile.data object

the metadata of P_new will be used join.
}}

