% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{add_metadata}
\alias{add_metadata}
\alias{add_metadata.profile.data}
\title{Add metadata}
\usage{
add_metadata(P, ...)

\method{add_metadata}{profile.data}(P, metadata, join_columns, ...)
}
\arguments{
\item{P}{profile.data}

\item{...}{additional parameters}

\item{metadata}{data.frame with new metadata columns}

\item{join_columns}{columns on which to join}
}
\description{
Add metadata
}
\section{Methods (by class)}{
\itemize{
\item \code{profile.data}: Add metadata to profile.data object
}}

