% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate rows}
\usage{
aggregate(population, variables, grouping_variables, operation = "mean", ...)
}
\arguments{
\item{population}{population}

\item{variables}{variables}

\item{grouping_variables}{grouping_variables}

\item{operation}{operation}

\item{...}{Arguments to be passed to methods}
}
\value{
data.frame after aggregation
}
\description{
Aggregate rows
}

