% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{assess}
\alias{assess}
\title{Assess track quality.}
\usage{
assess(tracks, min_path_length, strata)
}
\arguments{
\item{tracks}{data frame with track objects}

\item{min_path_length}{minimum length of a valid track}

\item{strata}{column name of track index column}
}
\value{
valid_observation_time
}
\description{
Assess track quality.
}
\examples{
 data <- tibble::data_frame(
   Metadata_timePoint = c(1:5),
   Location_Center_X = c(1, 2, 3, 4, 5),
   Location_Center_Y = c(1, 1, 1, 1, 1),
   TrackObjects_Label = c(rep(1, 5))
 )
 strata <- 'TrackObjects_Label'
 data <- dplyr::group_by_(data,strata)
 tracks <- track(data,strata)
 min_path_length <- 5
 trackQuality <- cytominer::assess(tracks,min_path_length,strata)
}
