% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CE_entropy_FS2_new}
\alias{CE_entropy_FS2_new}
\title{CE_entropy_FS2_new}
\usage{
CE_entropy_FS2_new(A, mc)
}
\arguments{
\item{A}{dataset mxn (m features and n observations)}

\item{mc}{number of feature to select}
}
\value{
index of features that were selected
}
\description{
Calculate the entropy of a matrix based on SVD with FS method
}
