% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CE_entropy_SR}
\alias{CE_entropy_SR}
\title{CE_entropy_SR}
\usage{
CE_entropy_SR(A)
}
\arguments{
\item{A}{dataset mxn (m features and n observations)}
}
\value{
entropy contribution vector
}
\description{
Calculate the entropy contribution of a matrix based on SVD
}
