% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{chemotactic_index}
\alias{chemotactic_index}
\title{Calculate the mean chemotactic index of a track object.}
\usage{
chemotactic_index(tracks)
}
\arguments{
\item{tracks}{data frame with track objects}
}
\value{
chemotactic_index
}
\description{
Calculate the mean chemotactic index of a track object.
}
\examples{
data <- tibble::data_frame(
  Metadata_timePoint = c(1:5),
  Location_Center_X = c(1, 2, 3, 4, 5),
  Location_Center_Y = c(1, 1, 1, 1, 1),
  TrackObjects_Label = c(rep(1, 5))
)
 tracks <- cytominer::displace(data,'TrackObjects_Label')
 chemotactic_index <-  cytominer::chemotactic_index(tracks)
}
