% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.mat.access.R
\name{col_meta}
\alias{col_meta}
\alias{col_meta.sim.mat}
\title{Access col metadata}
\usage{
col_meta(...)

\method{col_meta}{sim.mat}(S, ...)
}
\arguments{
\item{...}{object from which to extract col metadata}

\item{S}{sim.mat object}
}
\description{
Access col metadata
}
\section{Methods (by class)}{
\itemize{
\item \code{sim.mat}: Access col metadata
}}

