% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compute_similarity.R
\name{compute_similarity}
\alias{compute_similarity}
\title{Similarity between two vectors}
\usage{
compute_similarity(D1, D2, grouping_cols = NULL, method = "spearman")
}
\arguments{
\item{D1}{data.frame}

\item{D2}{same dimensions as \code{D1}}
}
\value{
Similarity between D1 and D2
}
\description{
Function to measure similarity between two vectors
}
\examples{
compute_similarity(c(1,2,3,4,5), c(1,2,4,3,5))
}

