% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity_within_group.R
\name{compute_similarity_within_group}
\alias{compute_similarity_within_group}
\title{Similarity between between all pairs of vectors within a group}
\usage{
compute_similarity_within_group(P, grp, method, ...)
}
\arguments{
\item{P}{profile.data}

\item{grp}{data.frame specifying list of grouping variables}

\item{method}{type of similarity method}

\item{...}{Arguments to be passed to methods}
}
\value{
list of sim.mats, one per group
}
\description{
Function to measure similarity between all pairs of vectors within a group
}

