% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_threshold.R
\name{correlation_threshold}
\alias{correlation_threshold}
\title{Select features using correlation threshold}
\usage{
correlation_threshold(variables, sample, cutoff = 0.9, method = "pearson")
}
\arguments{
\item{variables}{Vector of column names defining the used features.}

\item{sample}{Subpopulation used to calculate the distribution of the given variables.}

\item{cutoff}{Threshold is a variable between [0,1] that defines the minimum correlation of a selected feature. Default value 0.90.}

\item{method}{Defines the method used to calculate the correlation. Method is a character string indicating which correlation coefficient (or covariance) is used. One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}
}
\value{
List with excluded features.
}
\description{
Select features using correlation threshold
}
