% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{covariance}
\alias{covariance}
\title{Compute covariance matrix and vectorize.}
\usage{
covariance(population, variables)
}
\arguments{
\item{population}{tbl with grouping (metadata) and observation variables.}

\item{variables}{character vector specifying observation variables.}
}
\value{
data frame of 1 row comprising vectorized covariance matrix.
}
\description{
\code{covariance} computes the covariance matrix and vectorize it.
}
\examples{

population <- tibble::tibble(
  x = rnorm(30),
  y = rnorm(30),
  z = rnorm(30)
)

variables <- c("x", "y")

covariance(population, variables)
}
