% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpseedseq_cnt_prf.R
\docType{data}
\name{cpseedseq_cnt_prf}
\alias{cpseedseq_cnt_prf}
\title{Cell Painting data - cell counts only}
\format{A profile.data object with 1152 rows}
\usage{
cpseedseq_cnt_prf
}
\description{
Cell Painting data - cell counts only
}
\keyword{datasets}

