% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpseedseq_prf.R
\docType{data}
\name{cpseedseq_prf}
\alias{cpseedseq_prf}
\title{Cell Painting data}
\format{A profile.data object with 1152 rows}
\usage{
cpseedseq_prf
}
\description{
Cell Painting data
}
\keyword{datasets}

