% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytominer-package.R
\docType{package}
\name{cytominer-package}
\alias{cytominer}
\alias{cytominer-package}
\title{cytominer: Methods for Image-Based Cell Profiling}
\description{
`cytominer` is a suite of common functions used to process high-dimensional
readouts from high-throughput cell experiments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cytomining/cytominer}
  \item Report bugs at \url{https://github.com/cytomining/cytominer/issues}
}

}
\author{
\strong{Maintainer}: Shantanu Singh \email{shsingh@broadinstitute.org}

Authors:
\itemize{
  \item Tim Becker \email{tbecker@broadinstitute.org}
  \item Allen Goodman \email{allen.goodman@icloud.com}
  \item Claire McQuin \email{mcquin@broadinstitute.org}
  \item Mohammad Rohban \email{mrohban@broadinstitute.org}
}

}
\keyword{internal}
