% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{displace}
\alias{displace}
\title{Add spatial displacement per frame for each track object}
\usage{
displace(population, group_variable)
}
\arguments{
\item{population, }{data frame storing single cell data}

\item{group_variable, }{column name storing the track label}
}
\value{
displacement
}
\description{
Add spatial displacement per frame for each track object
}
\examples{
 data <- tibble::data_frame(
   Metadata_timePoint = c(1:5),
   Location_Center_X = c(1, 2, 3, 4, 5),
   Location_Center_Y = c(1, 1, 1, 1, 1),
   TrackObjects_Label = c(rep(1, 5))
 )
 tracks <- cytominer::displace(data,'TrackObjects_Label')
}
