% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_subpopulations.R
\name{extract_subpopulations}
\alias{extract_subpopulations}
\title{extract the subpopulations enriched/de-enriched in a given single cell population of a treatment w.r.t the control}
\usage{
extract_subpopulations(population_treatment, population_control, k)
}
\arguments{
\item{population_treatment}{single cell data for a given treatment, including just the features}

\item{population_control}{the same matrix for the negative control}

\item{k}{number of subpopulations}
}
\value{
list containing subpop. signatures and two histograms showing freq. of each subpop. in treatment and control.
}
\description{
extract the subpopulations enriched/de-enriched in a given single cell population of a treatment w.r.t the control
}

