% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/profile.data.access.R
\name{feat}
\alias{feat}
\alias{feat.profile.data}
\title{Access featdata}
\usage{
feat(...)

\method{feat}{profile.data}(P, ...)
}
\arguments{
\item{...}{}

\item{P}{profile.data object}
}
\description{
Access featdata
}
\section{Methods (by class)}{
\itemize{
\item \code{profile.data}: 
}}

