% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixture_correlation_threshold_intensities.R
\docType{data}
\name{fixture_correlation_threshold_intensities}
\alias{fixture_correlation_threshold_intensities}
\title{Correlation thresholded intensities data}
\format{A data.frame object with 4959 rows}
\usage{
fixture_correlation_threshold_intensities
}
\description{
Correlation thresholded intensities data
}
\keyword{datasets}

