% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixture_generalized_log_intensities.R
\docType{data}
\name{fixture_generalized_log_intensities}
\alias{fixture_generalized_log_intensities}
\title{Generalized log transformed intensities data}
\format{A data.frame object with 4 rows}
\usage{
fixture_generalized_log_intensities
}
\description{
Generalized log transformed intensities data
}
\keyword{datasets}

