% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixture_intensities.R
\docType{data}
\name{fixture_intensities}
\alias{fixture_intensities}
\title{Intensities data}
\format{A data.frame object with 4959 rows}
\usage{
fixture_intensities
}
\description{
Intensities data
}
\keyword{datasets}

