% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixture_intensities_small.R
\docType{data}
\name{fixture_intensities_small}
\alias{fixture_intensities_small}
\title{Intensities data}
\format{A data.frame object with 4 rows}
\usage{
fixture_intensities_small
}
\description{
Intensities data
}
\keyword{datasets}

