% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixture_linearized_intensities.R
\docType{data}
\name{fixture_linearized_intensities}
\alias{fixture_linearized_intensities}
\title{Linearized intensities data}
\format{A data.frame object with 4959 rows}
\usage{
fixture_linearized_intensities
}
\description{
Linearized intensities data
}
\keyword{datasets}

