% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixture_robustized_intensities.R
\docType{data}
\name{fixture_robustized_intensities}
\alias{fixture_robustized_intensities}
\title{Robustized intensities data}
\format{A data.frame object with 4959 rows}
\usage{
fixture_robustized_intensities
}
\description{
Robustized intensities data
}
\keyword{datasets}

