% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixture_standardized_intensities.R
\docType{data}
\name{fixture_standardized_intensities}
\alias{fixture_standardized_intensities}
\title{Standardized intensities data}
\format{A data.frame object with 4959 rows}
\usage{
fixture_standardized_intensities
}
\description{
Standardized intensities data
}
\keyword{datasets}

