% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixture_variance_threshold_shapes.R
\docType{data}
\name{fixture_variance_threshold_shapes}
\alias{fixture_variance_threshold_shapes}
\title{Variance thresholded shapes data}
\format{A data.frame object with 4959 rows}
\usage{
fixture_variance_threshold_shapes
}
\description{
Variance thresholded shapes data
}
\keyword{datasets}

