% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim.metric.R
\name{format.sim.metric}
\alias{format.sim.metric}
\title{Format a sim.metric object}
\usage{
\method{format}{sim.metric}(s, ...)
}
\arguments{
\item{s}{sim.metric object}
}
\description{
Format a sim.metric object
}

