% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pair_query.R
\name{format_pair_query}
\alias{format_pair_query}
\title{Format query contained in data frame}
\usage{
format_pair_query(cmp, dict_column_names)
}
\arguments{
\item{cmp}{list containing two data.frames corresponding to the query pair}

\item{dict_column_names}{column names of the data.frame dictionary
that will be queried}
}
\value{
Formatted query
}
\description{
Format query contained in data frame
}

