% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{lifeTime}
\alias{lifeTime}
\title{Calculate lifeTime of a track object.}
\usage{
lifeTime(tracks)
}
\arguments{
\item{tracks}{data frame with track objects}
}
\value{
Calculate life time of each track object
}
\description{
Calculate lifeTime of a track object.
}
\examples{
 data <- tibble::data_frame(
   Metadata_timePoint = c(1:5),
   Location_Center_X = c(1, 2, 3, 4, 5),
   Location_Center_Y = c(1, 1, 1, 1, 1),
   TrackObjects_Label = c(rep(1, 5))
 )
 tracks <- cytominer::displace(data,'TrackObjects_Label')
 lifeTime <-  cytominer::lifeTime(tracks)
   
}
