% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearize.R
\name{linearize}
\alias{linearize}
\title{Linearize rows}
\usage{
linearize(population, variables, grouping_variables, sample,
  lower_quantile = 0.25, upper_quantile = 0.75)
}
\arguments{
\item{population}{population}

\item{variables}{variables}

\item{grouping_variables}{grouping_variables}

\item{sample}{sample}

\item{lower_quantile}{lower_quantile}

\item{upper_quantile}{upper_quantile}
}
\value{
object after normalization
}
\description{
Linearize rows
}

