% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_outlier_rows.R
\name{mark_outlier_rows}
\alias{mark_outlier_rows}
\title{Mark outlier rows.}
\usage{
mark_outlier_rows(
  population,
  variables,
  strata,
  sample,
  operation = "svd+iqr",
  ...
)
}
\arguments{
\item{population}{tbl with grouping (metadata) and observation variables.}

\item{variables}{character vector specifying observation variables.}

\item{strata}{character vector specifying grouping variables for grouping
prior to outlier removal.}

\item{sample}{tbl containing sample that is used by outlier removal methods
to estimate parameters. \code{sample} has same structure as
\code{population}. Typically, \code{sample} corresponds to controls in the
experiment.}

\item{operation}{optional character string specifying method for outlier
removal. There is currently only one option (\code{"svd_iqr"}).}

\item{...}{arguments passed to outlier removal operation.}
}
\value{
\code{population} with an extra column \code{is_outlier}.
}
\description{
\code{mark_outlier_rows} drops outlier rows.
}
\examples{
suppressMessages(suppressWarnings(library(magrittr)))
population <- tibble::tibble(
  Metadata_group = c(
    "control", "control", "control", "control",
    "experiment", "experiment", "experiment", "experiment"
  ),
  Metadata_batch = c("a", "a", "b", "b", "a", "a", "b", "b"),
  AreaShape_Area = c(100, 12, 15, 16, 8, 8, 7, 7),
  AreaShape_Eccentricity = c(1, 5, 2, 1, 3, 8, 70, 4)
)
variables <- c("AreaShape_Area", "AreaShape_Eccentricity")
strata <- c("Metadata_batch")
sample <- population \%>\% dplyr::filter(Metadata_group == "control")
cytominer::mark_outlier_rows(population, variables, strata,
  sample,
  operation = "svd_iqr"
)
}
