% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{mean_squared_displacement}
\alias{mean_squared_displacement}
\title{Calculate the mean squared displacement of a track object.}
\usage{
mean_squared_displacement(tracks, tau)
}
\arguments{
\item{tracks}{data frame with track objects}

\item{tau}{delta t}
}
\value{
mean_squared_displacement
}
\description{
Calculate the mean squared displacement of a track object.
}
\examples{
 data <- tibble::data_frame(
   Metadata_timePoint = c(1:5),
   Location_Center_X = c(1, 2, 3, 4, 5),
   Location_Center_Y = c(1, 1, 1, 1, 1),
   TrackObjects_Label = c(rep(1, 5))
 )
 tau <- 2
 tracks <- cytominer::displace(data,'TrackObjects_Label')
 mean_squared_displacement <-  cytominer::mean_squared_displacement(tracks,tau)
   
}
