% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merge_by_xid.R
\name{merge_by_xid}
\alias{merge_by_xid}
\title{Join metadata and featdata}
\usage{
merge_by_xid(metadata, featdata)
}
\arguments{
\item{metadata}{data.frame}

\item{featdata}{data.frame}
}
\description{
Join metadata and featdata
}

