% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/profile.data.access.R
\name{meta}
\alias{meta}
\alias{meta.profile.data}
\title{Access metadata}
\usage{
meta(...)

\method{meta}{profile.data}(P, ...)
}
\arguments{
\item{...}{}

\item{P}{profile.data object}
}
\description{
Access metadata
}
\section{Methods (by class)}{
\itemize{
\item \code{profile.data}: 
}}

