% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_filter_metadata.R
\name{post_filter_metadata}
\alias{post_filter_metadata}
\alias{post_filter_metadata.profile.data}
\title{Filter rows of featdata by joining with metadata}
\usage{
post_filter_metadata(...)

\method{post_filter_metadata}{profile.data}(P, ...)
}
\arguments{
\item{...}{Arguments to be passed to methods}

\item{P}{profile.data}
}
\value{
profile.data, after filtering featdata
}
\description{
Used after metadata has been filtered
}
\section{Methods (by class)}{
\itemize{
\item \code{profile.data}: 
}}

