% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_metadata.R
\name{process_metadata}
\alias{process_metadata}
\alias{process_metadata.profile.data}
\title{Process metadata}
\usage{
process_metadata(P, strip_cellprofiler_db_tags = T)

\method{process_metadata}{profile.data}(P, strip_cellprofiler_db_tags = T)
}
\arguments{
\item{P}{profile.data}

\item{strip_cellprofiler_db_tags}{If True, drop the leading characters
in the metadata columns names that originate from the per_image table
of the CellProfiler output}
}
\description{
Process metadata
}
\section{Methods (by class)}{
\itemize{
\item \code{profile.data}: Process metadata for profile.data object
}}

