% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/profile.data.R
\name{profile.data}
\alias{profile.data}
\title{Constructor for profile.data S3 class}
\usage{
\method{profile}{data}(cf)
}
\arguments{
\item{cf}{YAML configuration file}
}
\value{
profile.data object
}
\description{
Constructor for profile.data S3 class
}

