% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query_n.R
\name{query_n}
\alias{query_n}
\alias{query_n.sim.mat}
\title{Query sim.mat - multiple queries are allowed}
\usage{
query_n(S, query_frame, ...)

\method{query_n}{sim.mat}(S, query_frame, return_all_cols = F, ...)
}
\arguments{
\item{S}{sim.mat object}

\item{query_frame}{data.frame with query in each row}

\item{...}{additional parameters}

\item{return_all_cols}{If True, returns all columns of the query result,
else returns only the columns that were present in the query}
}
\value{
data.frame with query result
}
\description{
This is a wrapper for query() that accepts multiple rows in the query frame
}
\section{Methods (by class)}{
\itemize{
\item \code{sim.mat}: Query sim.mat object - multiple queries are allowed
}}

