% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query_set.R
\name{query_set}
\alias{query_set}
\alias{query_set.list}
\title{Query list of sim.mats}
\usage{
query_set(S_l, query_frame, ...)

\method{query_set}{list}(S_l, query_frame, return_all_cols = F, ...)
}
\arguments{
\item{S_l}{list of sim.mats}

\item{query_frame}{data.frame with query in each row}

\item{...}{additional parameters}

\item{return_all_cols}{If True, returns all columns of the query result,
else returns only the columns that were present in the query}
}
\value{
data.frame with query result
}
\description{
This is a wrapper for query() that accepts multiple rows in the query frame
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: Query set of sim.mats
}}

