% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/replace_metadata.R
\name{replace_metadata}
\alias{replace_metadata}
\alias{replace_metadata.profile.data}
\title{Replace metadata}
\usage{
replace_metadata(P, ...)

\method{replace_metadata}{profile.data}(P, meta_new, key_cols, ...)
}
\arguments{
\item{P}{profile.data}

\item{...}{additional parameters}

\item{meta_new}{data.frame with new metadata columns}

\item{key_cols}{key columns on which to join}
}
\description{
Replace metadata
}
\section{Methods (by class)}{
\itemize{
\item \code{profile.data}: Replace metadata in a profile.data object
}}

