% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_correlation.R
\name{replicate_correlation}
\alias{replicate_correlation}
\title{Replicate correlation of variables}
\usage{
replicate_correlation(sample, variables, strata, replicates,
  replicate_by = NULL, split_by = NULL, cores = NULL)
}
\arguments{
\item{sample}{Population used to estimate the correlation.}

\item{variables}{Vector of column names defining the used features.}

\item{strata}{Vector of grouping variable.}

\item{replicates}{...}

\item{replicate_by}{... Default = NULL.}

\item{split_by}{. Default = NULL.}

\item{cores}{Number of cpu cores used for parallel computing using doParallel. Default cores = NULL.}
}
\value{
data.frame of variable quality measurements
}
\description{
Replicate correlation of variables
}
