% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rotate_platemap.R
\name{rotate_platemap}
\alias{rotate_platemap}
\alias{rotate_platemap.profile.data}
\title{Rotate the platemap for a 384-well plate}
\usage{
rotate_platemap(...)

\method{rotate_platemap}{profile.data}(P, ...)
}
\arguments{
\item{P}{profile.data generated from a 384-well plate experiment
and containing 'Well' in the metdata}
}
\value{
profile.data with platemap rotate by 180 degrees
}
\description{
Rotate the platemap for a 384-well plate
}
\section{Methods (by class)}{
\itemize{
\item \code{profile.data}: 
}}

