% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_dplyr.R
\name{scale_dplyr}
\alias{scale_dplyr}
\title{Scale and center a data.frame or a dplyr::tbl_sql}
\usage{
scale_dplyr(x, center, scale, vars)
}
\arguments{
\item{x}{data.frame or dplyr::tbl_sql}

\item{center}{center}

\item{scale}{scale}

\item{vars}{vars}
}
\value{
data.frame or dplyr::tbl_sql after normalization
}
\description{
Scale and center a data.frame or a dplyr::tbl_sql
}

