% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{score_features_sv_entropy}
\alias{score_features_sv_entropy}
\title{score_features_sv_entropy}
\usage{
score_features_sv_entropy(data)
}
\arguments{
\item{data}{a matrix which represents the dataset; columns and rows correspond to features and observations, respectively.}
}
\value{
vector containing scores for all the features, in the same order as the columns of data are arranged
}
\description{
Scores each feature based on the difference of normalized SVs entropy of the data
with and without the feature; the higher the difference, the more informative the
feature would be.
}
