% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{sectorAnalysis}
\alias{sectorAnalysis}
\title{perform sector analysis and label each track according to its direction of movement.}
\usage{
sectorAnalysis(tracks)
}
\arguments{
\item{tracks}{data frame with track objects}
}
\value{
sector
}
\description{
perform sector analysis and label each track according to its direction of movement.
}
\examples{
 data <- tibble::data_frame(
   Metadata_timePoint = c(1:5),
   Location_Center_X = c(1, 2, 3, 4, 5),
   Location_Center_Y = c(1, 1, 1, 1, 1),
   TrackObjects_Label = c(rep(1, 5))
 )
 tracks <- cytominer::displace(data,'TrackObjects_Label')
 sectorAnalysis <-  cytominer::sectorAnalysis(tracks)
   
}
