% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim.mat.R
\name{sim.mat}
\alias{sim.mat}
\title{Constructor for sim.mat S3 class}
\usage{
sim.mat(smat, row_meta, col_meta)
}
\arguments{
\item{cf}{YAML configuration file}

\item{use_csv}{If TRUE, force loading of CSV file even if binary exists}
}
\value{
profile.data object
}
\description{
Constructor for sim.mat S3 class
}

