% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.metric.R
\name{sim.metric}
\alias{sim.metric}
\title{Constructor for sim.metric S3 class}
\usage{
sim.metric(sim_metric_df)
}
\arguments{
\item{sim_metric_df}{data.frame describing the similarity metric}
}
\value{
sim.metric object
}
\description{
Constructor for sim.metric S3 class
}

