% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{singular_value_entropy}
\alias{singular_value_entropy}
\title{singular_value_entropy}
\usage{
singular_value_entropy(A)
}
\arguments{
\item{A}{a matrix of any arbitrary size}
}
\value{
entropy of the normalized singular values using log base 10
}
\description{
Calculate the entropy of a matrix singular values in the SVD decomposition, when their summation is normalized to one.
}
