% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim.mat.access.R
\name{smat}
\alias{smat}
\alias{smat.sim.mat}
\title{Access similarity matrix}
\usage{
smat(...)

\method{smat}{sim.mat}(S, ...)
}
\arguments{
\item{...}{object from which to extract similarity matrix}

\item{S}{sim.mat object}
}
\description{
Access similarity matrix
}
\section{Methods (by class)}{
\itemize{
\item \code{sim.mat}: 
}}

