% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\title{Summarize data}
\usage{
summarize(population, variables, operation, ...)
}
\arguments{
\item{population}{...}

\item{variables}{...}

\item{operation}{...}

\item{...}{arguments passed to summarization operation}
}
\value{
summarized data
}
\description{
Summarize data
}
