% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{track}
\alias{track}
\title{Compute track statistics}
\usage{
track(population, grouping_variable)
}
\arguments{
\item{population, }{single cell data}

\item{grouping_variable, }{column name storing the track label}
}
\value{
track
}
\description{
Compute track statistics
}
\examples{
 data <- tibble::data_frame(
   Metadata_timePoint = c(1:5),
   Location_Center_X = c(1, 2, 3, 4, 5),
   Location_Center_Y = c(1, 1, 1, 1, 1),
   TrackObjects_Label = c(rep(1, 5))
 )
 data <- dplyr::group_by_(data,'TrackObjects_Label')
 tracks <- track(data,'TrackObjects_Label')
}
