% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{trajectory}
\alias{trajectory}
\title{Compute trajectory statistics}
\usage{
trajectory(population, ...)
}
\arguments{
\item{population}{...}
}
\value{
trajectory statistics
}
\description{
Compute trajectory statistics
}
