% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{Transform rows of a data.frame}
\usage{
transform(population, sample = NULL, operation = "generalized_log", ...)
}
\arguments{
\item{population}{population}

\item{sample}{sample}

\item{operation}{operation}

\item{...}{Arguments to be passed to methods}
}
\value{
data.frame after transformation
}
\description{
Transform rows of a data.frame
}

