% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{validateTracks}
\alias{validateTracks}
\title{Identify valid tracks. Valid tracks are defined as tracks with a life time longer then a predifined value.}
\usage{
validateTracks(tracks, minPathLength)
}
\arguments{
\item{tracks}{data frame with track objects}

\item{minPathLength}{minimum length of a valid track}
}
\value{
validObservationTime
}
\description{
Identify valid tracks. Valid tracks are defined as tracks with a life time longer then a predifined value.
}
\examples{
 data <- tibble::data_frame(
   Metadata_timePoint = c(1:5),
   Location_Center_X = c(1, 2, 3, 4, 5),
   Location_Center_Y = c(1, 1, 1, 1, 1),
   TrackObjects_Label = c(rep(1, 5))
 )
 data <- dplyr::group_by_(data,'TrackObjects_Label')
 tracks <- track(data,'TrackObjects_Label')
 minPathLength <- 5
 validateTracks <-   validateTracks(tracks, minPathLength)
}
