% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance.R
\name{variable_importance}
\alias{variable_importance}
\title{Measure variable importance}
\usage{
variable_importance(sample, variables, operation = "replicate_correlation",
  ...)
}
\arguments{
\item{sample}{Poplation used to estimate the importance of each variable.}

\item{variables}{Vector of column names defining the used features.}

\item{operation}{The operation "replicate_correlation" can be chosen. Default operation = "replicate_correlation".}

\item{...}{arguments passed to variable importance operation}
}
\value{
variable importance measures
}
\description{
Measure variable importance
}
