% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_threshold.R
\name{variance_threshold}
\alias{variance_threshold}
\title{Select features of a data.frame using variance threshold}
\usage{
variance_threshold(population, ...)
}
\arguments{
\item{population}{population}

\item{...}{Arguments to be passed to methods}
}
\value{
data.frame after feature selection
}
\description{
Select features of a data.frame using variance threshold
}

