% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiten.R
\name{whiten}
\alias{whiten}
\title{variables <- c("Intensity_DNA", "Texture_DNA")
whiten(population, variables, sample)}
\usage{
whiten(population, variables, sample, regularization_param = 1)
}
\description{
variables <- c("Intensity_DNA", "Texture_DNA")
whiten(population, variables, sample)
}
