% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate data}
\usage{
aggregate(population, variables, strata, operation = "mean", ...)
}
\arguments{
\item{population}{...}

\item{variables}{...}

\item{strata}{...}

\item{operation}{...}

\item{...}{arguments passed to aggregation operation}
}
\value{
aggregated data
}
\description{
Aggregate data
}

