% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate data based on given grouping.}
\usage{
aggregate(population, variables, strata, operation = "mean", ...)
}
\arguments{
\item{population}{tbl with grouping (metadata) and observation variables.}

\item{variables}{character vector specifying observation variables.}

\item{strata}{character vector specifying grouping variables for aggregation.}

\item{operation}{optional character string specifying method for aggregation. This must be one of the strings \code{"mean"}, \code{"median"}, \code{"mean+sd"}, or \code{"cov"}.}

\item{...}{optional arguments passed to aggregation operation : \code{random_projection} is an optional matrix of size (# features * (#features + 1)/2, k); if operation is set to \code{"cov"}, this matrix would be used to project the resulting cov profile along k directions specified in the columns.}
}
\value{
aggregated data of the same class as \code{population}.
}
\description{
\code{aggregate} aggregates data based on the specified aggregation method.
}
\examples{
population <- tibble::data_frame(
   Metadata_group = c("control", "control", "control", "control",
                      "experiment", "experiment", "experiment", "experiment"),
   Metadata_batch = c("a", "a", "b", "b", "a", "a", "b", "b"),
   AreaShape_Area = c(10, 12, 15, 16, 8, 8, 7, 7)
 )
variables <- c("AreaShape_Area")
strata <- c("Metadata_group", "Metadata_batch")
aggregate(population, variables, strata, operation = "mean")

}
