% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{assess}
\alias{assess}
\title{Assess track quality.}
\usage{
assess(tracks, minPathLength, trackLabel)
}
\arguments{
\item{tracks}{data frame with track objects}

\item{minPathLength}{minimum length of a valid track}

\item{trackLabel}{column name of track index column}
}
\value{
validObservationTime
}
\description{
Assess track quality.
}
\examples{
 data <- tibble::data_frame(
   Metadata_timePoint = c(1:5),
   Location_Center_X = c(1, 2, 3, 4, 5),
   Location_Center_Y = c(1, 1, 1, 1, 1),
   TrackObjects_Label = c(rep(1, 5))
 )
 trackLabel <- 'TrackObjects_Label'
 data <- dplyr::group_by_(data,trackLabel)
 tracks <- track(data,trackLabel)
 minPathLength <- 5
 trackQuality <- assess(tracks,minPathLength,trackLabel)
}
