% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compute_similarity.R
\name{compute_similarity}
\alias{compute_similarity}
\title{Similarity between two vectors}
\usage{
compute_similarity(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector of the same length as x}
}
\value{
Similarity between x and y
}
\description{
Function to measure similarity between two vectors
}
\examples{
compute_similarity(c(1,2,3,4,5), c(1,2,4,3,5))
}

