% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_threshold.R
\name{correlation_threshold}
\alias{correlation_threshold}
\title{Select features using correlation threshold}
\usage{
correlation_threshold(variables, sample, cutoff = 0.9, method = "pearson")
}
\arguments{
\item{variables}{...}

\item{sample}{...}

\item{cutoff}{...}

\item{method}{...}
}
\value{
Excluded variables
}
\description{
Select features using correlation threshold
}
