% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_threshold.R
\name{correlation_threshold}
\alias{correlation_threshold}
\title{Select features using correlation threshold}
\usage{
correlation_threshold(population, variables, sample, cutoff = 0.9,
  method = "pearson", verbose = FALSE)
}
\arguments{
\item{population}{population}

\item{variables}{variables}

\item{sample}{sample}

\item{cutoff}{cutoff}

\item{method}{method}

\item{verbose}{verbose}
}
\value{
object after feature selection
}
\description{
Select features using correlation threshold
}

