% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_na_rows.R
\name{count_na_rows}
\alias{count_na_rows}
\title{Frequency of NAs per variable}
\usage{
count_na_rows(population, variables)
}
\arguments{
\item{population}{...}

\item{variables}{...}
}
\value{
data.frame with frequency of NAs per variable
}
\description{
Frequency of NAs per variable
}
