% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na_columns.R
\name{drop_na_columns}
\alias{drop_na_columns}
\title{Select features without NA values}
\usage{
drop_na_columns(population, variables)
}
\arguments{
\item{population}{...}

\item{variables}{...}
}
\value{
Excluded variables
}
\description{
Select features without NA values
}

