% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na_rows.R
\name{drop_na_rows}
\alias{drop_na_rows}
\title{Drop rows that are NA in all variables}
\usage{
drop_na_rows(population, variables)
}
\arguments{
\item{population}{...}

\item{variables}{...}
}
\value{
data without rows with NA in all variables
}
\description{
Drop rows that are NA in all variables
}

