% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_feature_selection.R
\name{entropy_feature_selection}
\alias{entropy_feature_selection}
\title{entropy based feature selection}
\usage{
entropy_feature_selection(x)
}
\arguments{
\item{x}{data matrix}
}
\description{
entropy based feature selection
}
