% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_subpopulations.R
\name{extract_subpopulations}
\alias{extract_subpopulations}
\title{Extract subpopulations.}
\usage{
extract_subpopulations(population, reference, variables, k)
}
\arguments{
\item{population}{tbl with grouping (metadata) and observation variables.}

\item{reference}{tbl with grouping (metadata) and observation variables. Columns of \code{population} and \code{reference} should be identical.}

\item{variables}{character vector specifying observation variables.}

\item{k}{scalar specifying number of subpopulations}
}
\value{
list containing subpopulation signatures (\code{subpop_centers}), two
histograms specifying frequency of each subpopulation in population and
reference (\code{subpop_profiles}), and cluster prediction and distance to
the predicted cluster for all input data (\code{treatment_clusters} and
\code{ctrl_clusters}).
}
\description{
\code{extract_subpopulations} extracts the subpopulations enriched/de-enriched in a given set w.r.t a reference set
}
\examples{
population <- tibble::data_frame(
   Metadata_group = c("control", "control", "control", "control",
                      "experiment", "experiment", "experiment", "experiment"),
   Metadata_batch = c("a", "a", "b", "b", "a", "a", "b", "b"),
   AreaShape_Area = c(10, 12, NA, 16, 8, 8, 7, 7),
   AreaShape_Length = c(2, 3, NA, NA, 4, 5, 1, 5)
)
variables <- c('AreaShape_Area','AreaShape_Length')
population_trt <-  dplyr::filter(population, Metadata_group == "experiment")
population_ctrl <- dplyr::filter(population, Metadata_group == "control")
extract_subpopulations(
   population = population_trt,
   reference = population_ctrl,
   variables = variables,
   k = 3
)

}
