% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_subpopulations.R
\name{extract_subpopulations}
\alias{extract_subpopulations}
\title{extract the subpopulations enriched/de-enriched in a given single cell population of a treatment w.r.t the control}
\usage{
extract_subpopulations(population_treatment, population_control, feats, k)
}
\arguments{
\item{population_treatment}{data frame which contains single cell data for a given treatment (cells are arranged in rows and measurements in columns)}

\item{population_control}{the same data frame (with the same column names, but possibly with different number of rows) for the negative control}

\item{feats}{a vector containing the feature names}

\item{k}{a scalar which is the number of subpopulations}
}
\value{
list containing subpop. signatures (subpop_centers), two histograms showing freq. of each subpop. in treatment and control (subpop_profiles), and cluster prediction and distance to the predicted cluster for all input data (treatment_clusters and ctrl_clusters).
}
\description{
extract the subpopulations enriched/de-enriched in a given single cell population of a treatment w.r.t the control
}
\examples{
population <- tibble::data_frame(
   Metadata_group = c("control", "control", "control", "control",
                      "experiment", "experiment", "experiment", "experiment"),
   Metadata_batch = c("a", "a", "b", "b", "a", "a", "b", "b"),
   AreaShape_Area = c(10, 12, NA, 16, 8, 8, 7, 7),
   AreaShape_Length = c(2, 3, NA, NA, 4, 5, 1, 5)
)
variables <- c('AreaShape_Area','AreaShape_Length')
population_trt <-  dplyr::filter(population, Metadata_group == "experiment")
population_ctrl <- dplyr::filter(population, Metadata_group == "control")
extract_subpopulations(
   population_treatment = population_trt,
   population_control = population_ctrl,
   feats = variables,
   k = 3
)

}
