% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_subpopulations.R
\name{extract_subpopulations}
\alias{extract_subpopulations}
\title{extract the subpopulations enriched/de-enriched in a given single cell population of a treatment w.r.t the control}
\usage{
extract_subpopulations(population_treatment, population_control, feats, k)
}
\arguments{
\item{population_treatment}{data frame which contains single cell data for a given treatment (cells are arranged in rows and measurements in columns)}

\item{population_control}{the same data frame (with the same column names, but possibly with different number of rows) for the negative control}

\item{feats}{a vector containing the feature names}

\item{k}{a scalar which is the number of subpopulations}
}
\value{
list containing subpop. signatures (subpop_centers), two histograms showing freq. of each subpop. in treatment and control (subpop_profiles), and cluster prediction and distance to the predicted cluster for all input data (treatment_clusters and ctrl_clusters).
}
\description{
extract the subpopulations enriched/de-enriched in a given single cell population of a treatment w.r.t the control
}

