% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_subpopulations.R
\name{extract_subpopulations}
\alias{extract_subpopulations}
\title{extract the subpopulations enriched/de-enriched in a given single cell population of a treatment w.r.t the control}
\usage{
extract_subpopulations(population_treatment, population_control, feats, k)
}
\arguments{
\item{population_treatment}{data frame whose rows are containing single cell data for a given treatment}

\item{population_control}{the same data frame (with the same column names) for the negative control}

\item{feats}{a vector containing the feature names}

\item{k}{a scalar which is the number of subpopulations}
}
\value{
list containing subpop. signatures and two histograms showing freq. of each subpop. in treatment and control.
}
\description{
extract the subpopulations enriched/de-enriched in a given single cell population of a treatment w.r.t the control
}

