% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.data.access.R
\name{full}
\alias{full}
\title{Access merged featdata and metadata}
\usage{
full(...)
}
\arguments{
\item{...}{object from which to extract merged featdata and metadata}
}
\description{
Access merged featdata and metadata
}

