% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_log.R
\name{generalized_log}
\alias{generalized_log}
\title{Generalized log transform data}
\usage{
generalized_log(population, variables, offset = 1)
}
\arguments{
\item{population}{Data frame with observation and grouping variables (metadata).}

\item{variables}{Vector of column names defining the used features.}

\item{offset}{Offset parameter for the log transformation. Default offset = 1.}
}
\value{
Data frame with generalized log transformed values.
}
\description{
Each variable is log transformed using x = log( (x + sqrt(x ^ 2 + offset ^ 2 )) / 2 ).
}
