% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_log.R
\name{generalized_log}
\alias{generalized_log}
\title{Generalized log of rows of a data.frame}
\usage{
generalized_log(population, variables, c = 1)
}
\arguments{
\item{population}{population}

\item{variables}{variables}

\item{c}{shift}
}
\value{
data.frame after transformation
}
\description{
Generalized log of rows of a data.frame
}

