% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_log.R
\name{generalized_log}
\alias{generalized_log}
\title{Generalized log transform data}
\usage{
generalized_log(population, variables, c = 1)
}
\arguments{
\item{population}{...}

\item{variables}{...}

\item{c}{...}
}
\value{
generalized log transformed data
}
\description{
Generalized log transform data
}

