% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/husk.R
\name{husk}
\alias{husk}
\title{Husk data.}
\usage{
husk(
  population,
  variables,
  sample,
  remove_outliers = TRUE,
  regularization_param = 1e-06,
  remove_signal = TRUE,
  flatten_noise = TRUE
)
}
\arguments{
\item{population}{tbl with grouping (metadata) and observation variables.}

\item{variables}{character vector specifying observation variables.}

\item{sample}{tbl containing sample that is used by the method to estimate
husking parameters. \code{sample} has same structure as \code{population}.
Typically, \code{sample} corresponds to controls in the experiment.}

\item{remove_outliers}{optional boolean specifying whether to remove
outliers. Default is \code{TRUE}.}

\item{regularization_param}{optional parameter used in husking to offset
eigenvalues to avoid division by zero. Default is \code{1}.}

\item{remove_signal}{optional boolean specifying whether to husk the signal
instead of only scaling it down. Default is \code{TRUE}.}

\item{flatten_noise}{optional boolean specifying whether to flatten the noise
instead of scaling it up. Default is \code{TRUE}. The parameter is ignored
if \code{remove_signal} is \code{FALSE}.}
}
\value{
transformed data of the same class as \code{population}.
}
\description{
\code{husk} detects unwanted variation in the sample and removes it from the
population.
}
\examples{
population <- tibble::tibble(
  Metadata_Well = c("A01", "A02", "B01", "B02"),
  Intensity_DNA = c(8, 20, 12, 32),
  Texture_DNA = c(5, 2, 43, 13)
)
variables <- c("Intensity_DNA", "Texture_DNA")
husk(population, variables, population, 1, remove_signal = TRUE)
husk(population, variables, population, 1e-5, remove_signal = TRUE)
husk(population, variables, population, 1, remove_signal = FALSE)
husk(population, variables, population, 1e-5, remove_signal = FALSE)
}
