% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearize.R
\name{linearize}
\alias{linearize}
\title{Linearize rows of a data.frame}
\usage{
linearize(population, sample, lower_quantile = 0.01, upper_quantile = 0.99)
}
\arguments{
\item{population}{population}

\item{sample}{sample}

\item{lower_quantile}{lower_quantile}

\item{upper_quantile}{upper_quantile}
}
\value{
data.frame after normalization
}
\description{
Linearize rows of a data.frame
}

