% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_outlier_rows.R
\name{mark_outlier_rows}
\alias{mark_outlier_rows}
\title{Mark outlier rows.}
\usage{
mark_outlier_rows(
  population,
  variables,
  sample,
  method = "svd+iqr",
  outlier_col = "is_outlier",
  ...
)
}
\arguments{
\item{population}{tbl with grouping (metadata) and observation variables.}

\item{variables}{character vector specifying observation variables.}

\item{sample}{tbl containing sample that is used by outlier removal methods
to estimate parameters. \code{sample} has same structure as
\code{population}. Typically, \code{sample} corresponds to controls in the
experiment.}

\item{method}{optional character string specifying method for outlier
removal. There is currently only one option (\code{"svd_iqr"}).}

\item{outlier_col}{optional character string specifying the name for the
column that will indicate outliers (in the output).
Default \code{"is_outlier"}.}

\item{...}{arguments passed to outlier removal method.}
}
\value{
\code{population} with an extra column \code{is_outlier}.
}
\description{
\code{mark_outlier_rows} drops outlier rows.
}
\examples{
suppressMessages(suppressWarnings(library(magrittr)))
population <- tibble::tibble(
  Metadata_group = sample(c("a", "b"), 100, replace = TRUE),
  Metadata_type = sample(c("control", "trt"), 100, replace = TRUE),
  AreaShape_Area = c(rnorm(98), 20, 30),
  AreaShape_Eccentricity = rnorm(100)
)
variables <- c("AreaShape_Area", "AreaShape_Eccentricity")
sample <- population \%>\% dplyr::filter(Metadata_type == "control")
population_marked <-
  cytominer::mark_outlier_rows(
    population,
    variables,
    sample,
    method = "svd+iqr"
  )
population_marked \%>\%
  dplyr::group_by(is_outlier) \%>\%
  dplyr::sample_n(3)
}
