% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize data}
\usage{
normalize(population, variables, strata, sample, operation = "standardize",
  ...)
}
\arguments{
\item{population}{...}

\item{variables}{...}

\item{strata}{...}

\item{sample}{...}

\item{operation}{...}

\item{...}{arguments passed to normalization operation}
}
\value{
normalized data
}
\description{
Normalize data
}

