% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize rows of a data.frame}
\usage{
normalize(population, sample, operation = "standardize", ...)
}
\arguments{
\item{population}{population}

\item{sample}{sample}

\item{operation}{operation}

\item{...}{Arguments to be passed to methods}
}
\value{
data.frame after normalization
}
\description{
Normalize rows of a data.frame
}

