% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R, R/normalize.R
\name{normalize_helper}
\alias{normalize_helper}
\title{Normalize helper}
\usage{
normalize_helper(group, population, variables, sample, centering_function,
  scaling_function)

normalize_helper(group, population, variables, sample, centering_function,
  scaling_function)
}
\arguments{
\item{group}{group}

\item{population}{population}

\item{variables}{variables}

\item{sample}{sample}

\item{centering_function}{centering_function}

\item{scaling_function}{scaling_function}

\item{group}{group}

\item{population}{population}

\item{variables}{variables}

\item{sample}{sample}

\item{centering_function}{centering_function}

\item{scaling_function}{scaling_function}
}
\value{
object after normalization

object after normalization
}
\description{
Normalize helper

Normalize helper
}

