% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/profile.data.R
\name{profile.data}
\alias{profile.data}
\title{Constructor for profile.data S3 class}
\usage{
\method{profile}{data}(cf = NULL, use_csv = F, imetadata = NULL,
  ifeatdata = NULL)
}
\arguments{
\item{cf}{YAML configuration file}

\item{use_csv}{If TRUE, force loading of CSV file even if binary exists}
}
\value{
profile.data object
}
\description{
Constructor for profile.data S3 class
}

