% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\alias{query.sim.mat}
\title{Query sim.mat}
\usage{
query(S, query_frame, ...)

\method{query}{sim.mat}(S, query_frame, ...)
}
\arguments{
\item{S}{sim.mat object}

\item{query_frame}{data.frame with query in each row}

\item{...}{additional parameters}
}
\description{
Query sim.mat
}
\section{Methods (by class)}{
\itemize{
\item \code{sim.mat}: Query sim.mat object
}}

