% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_correlation.R
\name{replicate_correlation}
\alias{replicate_correlation}
\title{Replicate correlation of variables}
\usage{
replicate_correlation(sample, variables, strata, replicates,
  replicate_by = NULL, split_by = NULL, cores = NULL)
}
\arguments{
\item{sample}{...}

\item{variables}{...}

\item{strata}{...}

\item{replicates}{...}

\item{replicate_by}{...}

\item{split_by}{...}
}
\value{
data.frame of variable quality measurements
}
\description{
Replicate correlation of variables
}

