% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustize.R
\name{robustize}
\alias{robustize}
\title{Robustize rows}
\usage{
robustize(population, variables, sample)
}
\arguments{
\item{population}{population}

\item{variables}{variables}

\item{sample}{sample}
}
\value{
object after normalization
}
\description{
Robustize rows
}

