% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustize.R
\name{robustize}
\alias{robustize}
\title{Robustize rows of a data.frame}
\usage{
robustize(population, sample)
}
\arguments{
\item{population}{population}

\item{sample}{sample}
}
\value{
data.frame after normalization
}
\description{
Robustize rows of a data.frame
}

