% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{score_features_sv_entropy}
\alias{score_features_sv_entropy}
\title{score_features_sv_entropy}
\usage{
score_features_sv_entropy(data)
}
\arguments{
\item{data}{a dataset mxn (m features and n observations)}
}
\value{
entropy contribution vector
}
\description{
Calculate the entropy contribution of a matrix based on SVD
}
