% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Select columns}
\usage{
select(population, variables, operation = "variance_threshold", ...)
}
\arguments{
\item{population}{...}

\item{variables}{...}

\item{operation}{...}

\item{...}{arguments passed to selection operation}
}
\value{
feature selected data
}
\description{
Select columns
}

