% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Select columns}
\usage{
select(population, variables, operation = "variance_threshold", ...)
}
\arguments{
\item{population}{population}

\item{variables}{variables}

\item{operation}{operation}

\item{...}{Arguments to be passed to methods}
}
\value{
object after feature selection
}
\description{
Select columns
}

