% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Select observation variables.}
\usage{
select(population, variables, sample = NULL,
  operation = "variance_threshold", ...)
}
\arguments{
\item{population}{tbl with grouping (metadata) and observation variables.}

\item{variables}{character vector specifying observation variables.}

\item{sample}{tbl containing sample that is used by some variable selection methods. \code{sample} has same structure as \code{population}.}

\item{operation}{optional character string specifying method for variable selection. This must be one of the strings \code{"variance_threshold"}, \code{"correlation_threshold"}, \code{"drop_na_columns"}.}

\item{...}{arguments passed to selection operation.}
}
\value{
variable-selected data of the same class as \code{population}.
}
\description{
\code{select} selects observation variables based on the specified variable selection method.
}

