% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim.mat.R
\name{sim.mat}
\alias{sim.mat}
\title{Constructor for sim.mat S3 class}
\usage{
sim.mat(smat, row_meta, col_meta, metric)
}
\arguments{
\item{smat}{Similarity matrix}

\item{row_meta}{data.frame col metadata}

\item{col_meta}{data.frame col metadata}
}
\value{
sim.mat object
}
\description{
Constructor for sim.mat S3 class
}

