% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize rows of a data.frame}
\usage{
standardize(population, variables, sample)
}
\arguments{
\item{population}{population}

\item{variables}{variables}

\item{sample}{sample}
}
\value{
data.frame after normalization
}
\description{
Standardize rows of a data.frame
}

