% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize rows}
\usage{
standardize(population, variables, grouping_variables, sample)
}
\arguments{
\item{population}{population}

\item{variables}{variables}

\item{grouping_variables}{grouping_variables}

\item{sample}{sample}
}
\value{
object after normalization
}
\description{
Standardize rows
}

