% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{stratify}
\alias{stratify}
\title{Stratify operations.}
\usage{
stratify(population, sample, reducer, strata, ...)
}
\arguments{
\item{population}{tbl with grouping (metadata) and observation variables.}

\item{sample}{tbl with the same structure as \code{population}. This is
typically used by operations to estimate parameters.}

\item{reducer}{operation that is to applied in a stratified manner.}

\item{strata}{optional character vector specifying grouping variables for
stratification.}

\item{...}{arguments passed to operation.}
}
\value{
\code{population} with potentially extra columns.
}
\description{
\code{stratify} stratifies operations.
}
\examples{
suppressMessages(suppressWarnings(library(magrittr)))
population <- tibble::tibble(
  Metadata_group = sample(c("a", "b"), 100, replace = TRUE),
  Metadata_type = sample(c("control", "trt"), 100, replace = TRUE),
  AreaShape_Area = c(rnorm(98), 20, 30),
  AreaShape_Eccentricity = rnorm(100)
)
variables <- c("AreaShape_Area", "AreaShape_Eccentricity")
strata <- c("Metadata_group")
sample <- population \%>\% dplyr::filter(Metadata_type == "control")
population_marked <-
  cytominer::stratify(
    reducer = cytominer::mark_outlier_rows,
    method = "svd+iqr",
    population = population,
    variables = variables,
    sample = sample,
    strata = strata
  )
population_marked \%>\%
  dplyr::group_by(is_outlier) \%>\%
  dplyr::sample_n(3)
}
