% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{Transform observation variables.}
\usage{
transform(population, variables, operation = "generalized_log", ...)
}
\arguments{
\item{population}{tbl with grouping (metadata) and observation variables.}

\item{variables}{character vector specifying observation variables.}

\item{operation}{optional character string specifying method for transform. Currently, only \code{"generalized_log"} (default) is implemented.}

\item{...}{arguments passed to transformation operation.}

\item{population}{Data frame with observation and grouping variables (metadata).}
}
\value{
transformed data of the same class as \code{population}.
}
\description{
\code{transform} transforms observation variables based on the specified transformation method.
}
