% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{Transform data}
\usage{
transform(population, variables, operation = "generalized_log", ...)
}
\arguments{
\item{population}{Data frame with observation and grouping variables (metadata).}

\item{variables}{Vector of column names defining the used features.}

\item{operation}{One of 'generalized log'. Default  operation = "generalized_log".}

\item{...}{arguments passed to transformation operation}
}
\value{
transformed data
}
\description{
Method to transform the valus in the data frame. This is often helpfull 
if the distribution of features does not follow a gaussian distribution.
}
