% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{Transform data}
\usage{
transform(population, variables, operation = "generalized_log", ...)
}
\arguments{
\item{population}{...}

\item{variables}{...}

\item{operation}{...}

\item{...}{arguments passed to transformation operation}
}
\value{
transformed data
}
\description{
Transform data
}

