% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{Transform rows}
\usage{
transform(population, variables, operation = "generalized_log", ...)
}
\arguments{
\item{population}{population}

\item{variables}{variables}

\item{operation}{operation}

\item{...}{Arguments to be passed to methods}
}
\value{
object after transformation
}
\description{
Transform rows
}

