% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance.R
\name{variable_importance}
\alias{variable_importance}
\title{Measure variable importance.}
\usage{
variable_importance(sample, variables, operation = "replicate_correlation",
  ...)
}
\arguments{
\item{sample}{tbl containing sample used to estimate parameters.}

\item{variables}{character vector specifying observation variables.}

\item{operation}{optional character string specifying method for computing variable importance. Currently, only \code{"replicate_correlation"} (default) is implemented.}

\item{...}{arguments passed to variable importance operation.}
}
\value{
data frame containing variable importance measures.
}
\description{
\code{variable_importance} measures importance of variables based on specified methods.
}

