% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_threshold.R
\name{variance_threshold}
\alias{variance_threshold}
\title{Select features using variance threshold}
\usage{
variance_threshold(population, variables, sample, ...)
}
\arguments{
\item{population}{population}

\item{variables}{variables}

\item{sample}{sample}

\item{...}{Arguments to be passed to methods}
}
\value{
Excluded variables
}
\description{
Select features using variance threshold
}

