% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_threshold.R
\name{variance_threshold}
\alias{variance_threshold}
\title{Select features using variance threshold}
\usage{
variance_threshold(variables, sample)
}
\arguments{
\item{variables}{Vector of column names defining the used features.}

\item{sample}{Subpopulation used to calculate the distribution of the given variables.}
}
\value{
Excluded variables
}
\description{
Select features using variance threshold
}
